+++ 06/ht/package.json
@@ -4,9 +4,9 @@
   "description": "Сайт состоит из трех блоков.",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "cross-env NODE_ENV=development webpack --mode development --config webpack.config.js",
-    "build": "cross-env NODE_ENV=production webpack --mode production --config webpack.config.js"
+    "build": "cross-env NODE_ENV=production webpack --mode production --config webpack.config.js",
+    "watch": "cross-env NODE_ENV=development webpack --mode development --config webpack.config.js --watch"
   },
   "keywords": [
     "js",
