--- 06/ht/src/AboutInfo.js
@@ -1,9 +1,0 @@
-function showAboutContent(element) {
-  element.innerHTML = `
-    <div class="content_about">
-        <h2>Информация о себе</h2>
-    </div>
-    `;
-}
-
-module.exports = showAboutContent;
