--- 06/ht/src/AuthorInfo.js
@@ -1,9 +1,0 @@
-function showAuthorContent(element) {
-  element.innerHTML = `
-  <div class="content_author">
-      <h2>Dima Kedik</h2>
-  </div>
-  `;
-}
-
-module.exports = showAuthorContent;
