+++ 06/ht/src/GoogleAPI.js
@@ -1,4 +1,4 @@
-async function resultGoogleMap(pos) {
+async function displayGoogleMap(pos) {
   mapboxgl.accessToken = "pk.eyJ1IjoiZGltYTUwODMiLCJhIjoiY2w4YzhlYmQzMHgzYjQwcGJjdTd2eGJuayJ9.msmLRpckL1zuR5vmK3M_eA";
   const map = new mapboxgl.Map({
     container: "map",
@@ -9,4 +9,4 @@ async function resultGoogleMap(pos) {
   const marker2 = new mapboxgl.Marker({ color: "red", rotation: 0 }).setLngLat(pos).addTo(map);
 }
 
-module.exports = resultGoogleMap;
+module.exports =  displayGoogleMap;
